requirements.txt
hydra-core==1.3.2
kaggle==1.6.17
matplotlib==3.10.0
pytorch-lightning==2.5.0.post0
python-dotenv==1.0.1
timm==1.0.12
torch==2.5.1
torchvision==0.20.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
wandb==0.19.2
